fix: fixed rudder_logger un-initialized issue (#69)
* chore: reading writekey and dataplane url from .env

* fix: RudderLogger Un-initialized issue

* feat: update example app to contain multiple screens

* feat: added navigation observer to example app to fire screen calls on screen change

* chore: removed .env from gitignore

* chore: added .env file with invalid dataplane url and writekey

* chore: updated README of example app

* chore: minor refactoring in the example app

* fix: fixed issues in unit tests in the example app
