+++ packages/example/.env
@@ -1,0 +1,2 @@
+DATA_PLANE_URL=DUPLICATE_DATA_PLANE_URL
+WRITE_KEY=DUPLICATE_WRITE_KEY
