+++ packages/example/README.md
@@ -1,0 +1,3 @@
+## Pre-requisites to run the Example app:
+
+*  Open `.env` and replace the values of `DUPLICATE_WRITE_KEY` and `DUPLICATE_DATA_PLANE_URL` with the actual values form your Rudderstack Dashboard.
