+++ packages/example/test/widget_test.dart
@@ -7,12 +7,12 @@
 
 // ignore_for_file: test-imports
 import 'package:flutter_test/flutter_test.dart';
-import 'package:rudder_sdk_flutter_example/main.dart';
+import 'package:rudder_sdk_flutter_example/home_screen.dart';
 
 void main() {
   testWidgets('Verify Platform version', (WidgetTester tester) async {
     // Build our app and trigger a frame.
-    await tester.pumpWidget(const PlatformChannel());
+    await tester.pumpWidget(const HomeScreen());
     expect(find.text('Track'), findsOneWidget);
   });
 }
