+++ packages/plugins/rudder_plugin_interface/lib/src/rudder_logger.dart
@@ -21,7 +21,6 @@ class RudderLogger {
       l = NONE;
     }
     __logLevel = l;
-    // _logger =
   }
 
   static void logError(String message) {
